NOISSUE - Refactor Messaging Documentation To Add Message Broker Support (#160)
* Refactor messaging documentation and add MQTT broker support

The commit introduces changes to the messaging documentation, specifically in the "MQTT Broker" section. It adds information about Mainflux's support for MQTT protocol and its usage for message exchange. The commit also mentions the MQTT version supported by Mainflux and the underlying MQTT client library used.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Fix configuration documentation for MQTT and message brokers

The configuration documentation for MQTT and message brokers in Mainflux has been updated to provide clearer instructions on how to handle different MQTT and message brokers in the dev environment. The documentation now includes information on the two types of brokers used by Mainflux, namely `MQTT_BROKER` and `MESSAGE_BROKER`, and the options available for each. Specifically, it clarifies that `MQTT_BROKER` can be either `vernemq` or `nats`, while `MESSAGE_BROKER` can be `nats`.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Refactor MQTT section in messaging.md

The MQTT section in the messaging.md file has been refactored to improve clarity and accuracy. The changes include rephrasing the description of MQTT as a lightweight Publish/Subscribe messaging protocol used for connecting restricted devices in low bandwidth, high-latency, or unreliable networks. Additionally, the sentence about the broker's responsibility for distributing messages to and from clients connected to the MQTT adapter has been truncated.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Rename mainflux to Mainflux

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Add Nats JS documentation

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Rename mainflux to Mainflux

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

---------

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>