NOISSUE - Add Rabbitmq Architectural Docs (#156)
* Refactor messaging documentation and add MQTT broker support

The commit introduces changes to the messaging documentation, specifically in the "MQTT Broker" section. It adds information about Mainflux's support for MQTT protocol and its usage for message exchange. The commit also mentions the MQTT version supported by Mainflux and the underlying MQTT client library used.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Add Nats JS documentation

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Rename mainflux to Mainflux

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Add rabbitmq architectural docs

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Fix changes after rebase

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

* Refactor messaging documentation

The messaging documentation has been refactored to update the name of the exchange used in Mainflux from `mainflux` to `messages`. This change ensures consistency and clarity in the documentation.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>

---------

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>
Co-authored-by: Dušan Borovčanin <dusan.borovcanin@mainflux.com>